aider: Added links to other submodules in the home page.
# Context:
USER: Please connect the home page to the other sub modules of this application
ASSISTANT: To connect the home page to the other submodules of your application, you can add links to the other pages in your `home.html` template. Here's how you can do it:

```html
inventory/templates/inventory/home.html
<<<<<<< ORIGINAL
<!DOCTYPE html>
<html>
<head>
    <title>Inventory Home</title>
</head>
<body>
    <h1>Welcome to the Inventory Home Page!</h1>
</body>
</html>
=======
<!DOCTYPE html>
<html>
<head>
    <title>Inventory Home</title>
</head>
<body>
    <h1>Welcome to the Inventory Home Page!</h1>
    <ul>
        <li><a href="{% url 'inventory:products' %}">Products</a></li>
        <li><a href="{% url 'inventory:stocks' %}">Stocks</a></li>
        <li><a href="{% url 'inventory:suppliers' %}">Suppliers</a></li>
        <li><a href="{% url 'inventory:purchase_orders' %}">Purchase Orders</a></li>
        <li><a href="{% url 'inventory:sales_orders' %}">Sales Orders</a></li>
    </ul>
</body>
</html>
>>>>>>> UPDATED
```

In the updated section, I've added a list of links to the other pages in your application. The `{% url %}` template tag is used to generate the URL for each page. The argument to the `{% url %}` tag is the name of the URL pattern you want to link to. The format is `'app_name:url_name'`.

Please note that this assumes that you have named your URL patterns in your `urls.py` file. If you haven't done this, you will need to add names to your URL patterns. For example:

```python
path('products/', views.ProductListView.as_view(), name='products'),
```

In this example, `'products'` is the name of the URL pattern.