aider: Fix issue with retrieving 'username' from POST data by using the `get` method instead of direct indexing.
# Context:
USER:   File "/Users/drakestapleton/Developer/crm/dcrm/website/views.py", line 13, in home
    username = request.POST['username']
               ~~~~~~~~~~~~^^^^^^^^^^^^
  File "/Users/drakestapleton/Developer/crm/crm/lib/python3.11/site-packages/django/utils/datastructures.py", line 86, in __getitem__
    raise MultiValueDictKeyError(key)
django.utils.datastructures.MultiValueDictKeyError: 'username'
ASSISTANT: The error message indicates that the key 'username' does not exist in the POST data of the request. This could happen if the form in your HTML does not have an input field with the name 'username', or if the form was submitted without filling that field.

To fix this, you can use the `get` method which will return `None` if the key does not exist, instead of raising an error. Here is how you can modify your code:

```python
website/views.py
<<<<<<< ORIGINAL
    username = request.POST['username']
=======
    username = request.POST.get('username')
>>>>>>> UPDATED
```

Please make sure that your form in the HTML file has an input field with the name attribute set to 'username'. If the user does not provide a username, this code will not raise an error, but the authentication will fail, as the username will be `None`.