aider: Add navbar to inventory application pages.
# Context:
USER: please add the nav-bar to the inventory application pages
ASSISTANT: To add the navbar to the inventory application pages, you need to include the navbar.html in each of the HTML templates of the inventory application.

Here's how you can do it:

```html
inventory/templates/inventory/home.html
<<<<<<< ORIGINAL
<!DOCTYPE html>
<html>
<head>
    <title>Inventory Home</title>
    <!-- Bootstrap CSS -->
    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css">
</head>
<body>
    <h1>Welcome to the Inventory Home Page!</h1>
=======
{% extends "base.html" %}
{% block content %}
    <h1>Welcome to the Inventory Home Page!</h1>
>>>>>>> UPDATED
```

You need to repeat this process for all the HTML templates in the inventory application.

Please note that this assumes that the navbar.html file is located in the templates directory of the website application and that the base.html file includes the navbar using `{% include "navbar.html" %}`. If the location of these files is different, you need to adjust the paths accordingly.