aider: Change the back button in the `StockDetailView` to redirect to the stock list page instead of the inventory home page.
# Context:
USER: edit the backbutton in the `StockDetailView`
ASSISTANT: To modify the back button in the `StockDetailView`, we need to edit the `inventory/templates/inventory/stock_detail.html` file. However, your request is a bit ambiguous. Could you please provide more details about how you want the back button to be modified? For example, do you want to change its appearance, its functionality, or the page it redirects to?
USER: make the button direct to the stock list file instead of home
ASSISTANT: To change the back button's redirection in the `StockDetailView` to the stock list, we need to modify the `inventory/templates/inventory/stock_detail.html` file.

Here is the change:

```html
inventory/templates/inventory/stock_detail.html
<<<<<<< ORIGINAL
<head>
    <a href="{% url 'inventory:inventory_home' %}" class="btn btn-secondary">Back</a>
    <title>Stock Detail</title>
=======
<head>
    <a href="{% url 'inventory:stocks' %}" class="btn btn-secondary">Back</a>
    <title>Stock Detail</title>
>>>>>>> UPDATED
```

This change will make the back button redirect to the stock list page instead of the inventory home page.