aider: Modified `generate_orders` view function to create `Order` and `Product` instances and save them to the database.
# Context:
USER: continue
ASSISTANT: To save the generated orders to the database, we need to modify the `generate_orders` view function to create `Order` and `Product` instances and save them to the database.

Here are the changes:

```python
accounting/views.py
<<<<<<< ORIGINAL
def generate_orders(request):
    order_generator = OrderGenerator(datetime.date.today() - datetime.timedelta(days=30), datetime.date.today())
    orders = order_generator.generate_orders(10)
    return JsonResponse(orders, safe=False)
=======
def generate_orders(request):
    order_generator = OrderGenerator(datetime.date.today() - datetime.timedelta(days=30), datetime.date.today())
    generated_orders = order_generator.generate_orders(10)
    for order in generated_orders:
        product_names = order['product_names'].split(', ')
        for product_name in product_names:
            product, created = Product.objects.get_or_create(name=product_name)
            Order.objects.create(customer_name=order['customer_name'], product=product, order_date=order['order_date'])
    return JsonResponse({"message": "Orders generated and saved to the database."})
>>>>>>> UPDATED
```

Please note that this code assumes that the `Order` model has a `customer_name` field, a `product` field (which is a foreign key to the `Product` model), and an `order_date` field. If the `Order` model does not have these fields, you need to modify the code accordingly. Also, this code does not handle the `total_price` field in the generated order. If you want to save the `total_price` to the database, you need to add a `total_price` field to the `Order` model and modify the code to save the `total_price`.