aider: Modify the `DATABASE_ROUTERS` setting in `dcrm/settings.py` to use the correct path for the `AccountingRouter` class.
# Context:
USER: please complete the settings file
ASSISTANT: To complete the settings file, I need more information about what you want to add or change. Here are some common things that might need to be added to a Django settings file:

1. Additional installed apps: If you've added new Django apps to your project, they need to be added to the `INSTALLED_APPS` list.
2. Middleware: If you're using any middleware, it needs to be added to the `MIDDLEWARE` list.
3.
^C KeyboardInterrupt
USER: I need you to finish setting up the mongodb database to be used with only the accounting module.
ASSISTANT: To set up the MongoDB database to be used only with the accounting module, we need to modify the `DATABASE_ROUTERS` setting in the `dcrm/settings.py` file.

Currently, it's set to `['Accounting.Router']`, but the correct path should be `['accounting.router.AccountingRouter']`. This is because Django expects the full Python import path to the router class.

Here is the change:

```python
dcrm/settings.py
<<<<<<< ORIGINAL
DATABASE_ROUTERS = ['Accounting.Router']
=======
DATABASE_ROUTERS = ['accounting.router.AccountingRouter']
>>>>>>> UPDATED
```

This change tells Django to use the `AccountingRouter` class in the `accounting/router.py` file to decide which database to use for each model. According to the logic in `AccountingRouter`, all models in the `accounting` app will use the `mongodb` database.