aider: Modified `OrderGeneratorForm` to allow the user to select dates in the past by rendering the `start_date` and `end_date` fields as HTML5 date input fields with a `max` attribute set to today's date.
# Context:
USER: great now make it so the user can select dates in the past
ASSISTANT: To allow the user to select dates in the past, we need to modify the `OrderGeneratorForm` in `accounting/forms.py`. We will set the `start_date` and `end_date` fields to use a `DateInput` widget and set the `max_date` to today's date.

```python
accounting/forms.py
<<<<<<< ORIGINAL
class OrderGeneratorForm(forms.Form):
    start_date = forms.DateField(widget=forms.SelectDateWidget())
    end_date = forms.DateField(widget=forms.SelectDateWidget())
    num_orders = forms.IntegerField(min_value=1)
=======
from django.forms.widgets import DateInput

class OrderGeneratorForm(forms.Form):
    start_date = forms.DateField(widget=DateInput(attrs={'type': 'date', 'max': datetime.date.today()}))
    end_date = forms.DateField(widget=DateInput(attrs={'type': 'date', 'max': datetime.date.today()}))
    num_orders = forms.IntegerField(min_value=1)
>>>>>>> UPDATED
```

This change will render the `start_date` and `end_date` fields as HTML5 date input fields, which allow the user to select a date from a date picker. The `max` attribute is set to today's date, so the user cannot select a date in the future.