aider: Modified the `orders.html` file to directly access the product's name and the order's quantity.
# Context:
USER: nothing is showing under the product and quanity coumns of my table
ASSISTANT: It seems like the issue might be in the way you're trying to access the product and quantity information in your template. Let's modify the `orders.html` file to correctly display the product and quantity information.

```html
accounting/templates/orders.html
<<<<<<< ORIGINAL
            <td>
                {% for product_name, quantity in order.product_quantities.items %}
                    {{ product_name }}: {{ quantity }}<br>
                {% endfor %}
            </td>
            <td>{{ order.quantity }}</td>
=======
            <td>{{ order.product.name }}</td>
            <td>{{ order.quantity }}</td>
>>>>>>> UPDATED
```

In the updated code, we're directly accessing the product's name and the order's quantity. Please note that this change assumes that each order has a single product. If an order can have multiple products, you'll need a different approach to display all the products and their quantities.