aider: Display multiple products and their quantities in the `orders.html` file by looping through all the products in an order.
# Context:
USER: Multiple products
ASSISTANT: If an order can have multiple products, we need to adjust the way we display the products and their quantities in the `orders.html` file.

```html
accounting/templates/orders.html
<<<<<<< ORIGINAL
            <td>{{ order.product.name }}</td>
            <td>{{ order.quantity }}</td>
=======
            <td>
                {% for product in order.products.all %}
                    {{ product.name }}<br>
                {% endfor %}
            </td>
            <td>
                {% for product in order.products.all %}
                    {{ product.quantity }}<br>
                {% endfor %}
            </td>
>>>>>>> UPDATED
```

In the updated code, we're looping through all the products in an order and displaying each product's name and quantity on a new line. This change assumes that the `Order` model has a many-to-many relationship with the `Product` model and that the `Product` model has a `quantity` field. If this is not the case, we'll need to adjust the models accordingly.